requirements.txt
django==5.2.8
django-rest-framework==0.1.0
djangorestframework-simplejwt==5.5.1
psycopg2==2.9.11